fix updating stat scripts. ;) another src's correcting.

+++ src/clserv.h
@@ -1,4 +1,4 @@
-/* $Id: clserv.h,v 1.5 2004/01/21 15:40:41 sisoft Exp $ */
+/* $Id: clserv.h,v 1.6 2004/02/13 22:29:01 sisoft Exp $ */
 #ifndef __CLSERV_H__
 #define __CLSERV_H__
 
@@ -27,13 +27,13 @@ typedef struct _cls_ln_t {
 	char *emsi;
 } cls_ln_t;
 
-extern int (*xsend_cb)(int sock,char *buf,int len);
+extern int (*xsend_cb)(int sock,char *buf,size_t len);
 
 extern int cls_conn(int type,char *port);
 extern void cls_close(int sock);
 extern void cls_shutd(int sock);
-extern int xsendto(int sock,char *buf,int len,struct sockaddr *to);
-extern int xsend(int sock,char *buf,int len);
-extern int xrecv(int sock,char *buf,int len,int wait);
+extern int xsendto(int sock,char *buf,size_t len,struct sockaddr *to);
+extern int xsend(int sock,char *buf,size_t len);
+extern int xrecv(int sock,char *buf,size_t len,int wait);
 
 #endif
